draws sentences but as squares??
