+++ client/include/sge/GraphicsShaders.h
@@ -17,8 +17,13 @@
 #include <fstream>
 #include <string>
 #include <vector>
+#include <map>
 #include <cmath>   // for sin() and other math functions
 #include <ctime>   // for time()
+
+// for freetype 
+#include <ft2build.h>
+#include FT_FREETYPE_H
 
 #include "sge/GraphicsConstants.h"
 
@@ -282,14 +287,24 @@ namespace sge {
 
     };
 
+    struct Character {
+        unsigned int TextureID;  // ID handle of the glyph texture
+        glm::ivec2   Size;       // Size of glyph
+        glm::ivec2   Bearing;    // Offset from baseline to left/top of glyph
+        unsigned int Advance;    // Offset to advance to next glyph
+    };
     class TextShaderProgram : public ShaderProgram {
     public:
         void initShaderProgram(const std::string &vertexShaderPath, const std::string &fragmentShaderPath);
+        void renderText(std::string text, float x, float y, float scale, glm::vec3 color);
     private:
         GLuint VAO;
         GLuint VBO;
 
         GLint projectionPos;
+
+        std::map<char, Character> Characters;
+        void loadFont();
 
     };
 
@@ -298,3 +313,5 @@ namespace sge {
     extern UIShaderProgram uiShaderProgram;
     extern TextShaderProgram textShaderProgram;
 }
+
+
